refactor: apis 폴더 자잘한 부분 리팩토링
