+++ src/apis/favor/FavorApi.ts
@@ -1,4 +1,5 @@
 import { axiosApi } from '@apis/axios'
+
 import {
   FavorAnswerRequest,
   FavorQuestionRequest,
