+++ src/pages/settingPage/LogOutBtn.tsx
@@ -9,7 +9,7 @@ import { useState } from 'react'
 import { useOutsideClick } from '@libs/hooks/useOutsideClick'
 import useSnackBar from '@libs/hooks/useSnackBar'
 import { useMutation } from '@tanstack/react-query'
-import { SettingApi } from '@apis/setting'
+import { SettingApi } from '@apis/setting/SettingApi'
 import { useNavigate } from 'react-router-dom'
 import { useRecoilValue } from 'recoil'
 import { authState } from '@libs/store/auth'
